% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{calculate_nonrandom_percent_on_heatmap_data}
\alias{calculate_nonrandom_percent_on_heatmap_data}
\title{calculate_nonrandom_percent_on_heatmap_data}
\usage{
calculate_nonrandom_percent_on_heatmap_data(
  heatmap.data,
  lrpd.data,
  upperPercentile = 0.95,
  d_cutoff = c(0)
)
}
\arguments{
\item{heatmap.data}{A result dataframe of simulate_population_remove_from_original_for_smaller_pops}

\item{lrpd.data}{A result dataframe of generate_large_randomlike_population_distribution - containing a Large Random-Like Population Distribution for each CM}

\item{upperPercentile}{Default=0.95 (95th percentile), the percentile of the LRPD to be taken as the MNRT per CM. For negative DRF, 1-upperPercentile will be taken as maximum instead.}

\item{d_cutoff}{Which dfr (Degree of Clumping) percentages to keep for the final MNRT}
}
\value{
An aggregated (mean) dataframe, with the columns of heatmap.data with and additional column for every CM in heatmap.data (column 3 and above), such that the name of the column
for a CM named 'name' will be called 'name.non.random.percent'. The results are the mean value (in the original column) and the mean
Non-Random Percent (in the new columns) per CSS + dfr pair (Current Sample Size and Degree of Clumping respectively).
This data can later be plotted in various ways to test the robustness of each CM.
}
\description{
For each Clumping Method (CM) of columns 3 and above in both heatmap.data and lrpd.data:
\itemize{
\item{}{Calculate the Minimum Non-Random Threshold (MNRT) - the upperPercentile percentile of the lrpd.data for the CM}
\item{}{Append a boolean column to the heatmap.data, such that 1 means a CM value over the MNRT}
}
The returned value will contain the Non-Random Percent (NRP) per Current Sample Size (CSS) and Degree of Clumping (dfr) pair.
}
