% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{calculate_values_example_data}
\alias{calculate_values_example_data}
\title{calculate_values_example_data}
\usage{
calculate_values_example_data(
  example_individual_location_data,
  funcs,
  check_clumping_method_valid = T
)
}
\arguments{
\item{example_individual_location_data}{intended for example data bird.location.data, but can be used for any dataframe with the following columns: POPULATION (character), X, Y.}

\item{funcs}{vector of function names. Will be tested for validity as a Clumping Method.}

\item{check_clumping_method_valid}{default=TRUE, check if clumping method is valid or not (via check_clumping_method_validity)}
}
\value{
A dataframe with the following columns: POPULATION, css (amount of individuals), and a column with the value of every CM on the population.
}
\description{
On example data bird.location.data, for each population, calculate the grade of each Clumping Method (CM) given by funcs.
}
