% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{check_clumping_method_validity}
\alias{check_clumping_method_validity}
\title{check_clumping_method_validity}
\usage{
check_clumping_method_validity(polys, cm_to_check, sim.type)
}
\arguments{
\item{polys}{List of sf polygons to check on the clumping methods}

\item{cm_to_check}{Clumping Method (CM) to check}

\item{sim.type}{simulation type parameter - to know which direction to check.}
}
\value{
A list with a success parameter (boolean), and a message of the error if success == FALSE.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{     ## If method is a valid CM
     list(success = T, message = '')
#> $success
#> [1] TRUE
#> 
#> $message
#> [1] ""
     
     ## If method is NOT a valid CM
     list(success = F, message = 'AN ERROR OCCURRED')
#> $success
#> [1] FALSE
#> 
#> $message
#> [1] "AN ERROR OCCURRED"
}\if{html}{\out{</div>}}
}
\description{
Given list of sf polygons and a function name, check if it works as a clumping method (CM):
\itemize{
\item{}{Exists in memory}
\item{}{Accepts parameters polys, dataframe, and n.for.calc}
\item{}{Return a numeric value}
\item{}{The numeric value is rising as clumpedness rises}
}
}
