% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{generate_clumped_population}
\alias{generate_clumped_population}
\title{generate_clumped_population}
\usage{
generate_clumped_population(
  n,
  polygons,
  n_for_clumping = 0,
  n_clumps = 0,
  max_dist_from_centroid_divisor = 2,
  min_dist_between_clumps_multiplier = 1.5
)
}
\arguments{
\item{n}{number of individuals in the population}

\item{polygons}{list of sf polygons in which to randomize the population}

\item{n_for_clumping}{used for calculating the expected distance for the clumping. Can cause closer clumpedness with n_for_clumping > n}

\item{n_clumps}{number of clumps wanted in the clumped populations. Default=0 uses the default function of floor(sqrt(n))}

\item{max_dist_from_centroid_divisor}{divisor of the expected distance for distances of each point from it's randomized centroid.
Higher number causes smaller distances between points in the same clump.}

\item{min_dist_between_clumps_multiplier}{the multiplier of the same distance for the distances between the centroids/clumps.}
}
\value{
A dataframe of the clumped population, with columns TAG (numeric, 1-n), X, Y, and clump.id (running index of the clump of the individual).
}
\description{
Generate a clumped population of n points within polygons in polys, return as a dataframe. Currently the resulting population is split into clumps of sqrt(n) points.
}
