% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{generate_large_randomlike_population_distribution}
\alias{generate_large_randomlike_population_distribution}
\title{generate_large_randomlike_population_distribution}
\usage{
generate_large_randomlike_population_distribution(
  polygons,
  function_names,
  helper_functions_to_export,
  sim.type = simulation_type$random.to.clumped,
  uniform.random.generation.type =
    uniform_random_generation_type$split.polygon.and.generate.separately,
  uniform.noise.ratio = 0.1,
  n.for.calc,
  n.for.clumping = 0,
  n.for.uniform = 0,
  initial_div = 2,
  max_pct_to_use = 20,
  percent_random_rise = 10,
  simulations_per_percent = 100,
  recalculate_poly = T,
  recalculate_poly_with_rect = F,
  plotResults = F,
  n_cores = 6
)
}
\arguments{
\item{polygons}{list of sf polygons where the populations can be generated.}

\item{function_names}{vector of function names (Clumping Methods, CMs), which are first tested by check_function_validity.}

\item{helper_functions_to_export}{vector of additional functions needed for the CMs to run. Tested only for existence in memory.}

\item{sim.type}{default = simulation_type$random.to.clumped, which part of the range to simulate.}

\item{uniform.random.generation.type}{default = uniform_random_generation_type$split.polygon.and.generate.separately, how to generate the populations in the uniform-random range (DFR -100 to 0), if relevant.}

\item{uniform.noise.ratio}{default = 0.1, noise for uniform function}

\item{n.for.calc}{sent into each of the function in function_names, but does not have to be used by each of these. Is meant}

\item{n.for.clumping}{similarly, the clumping function above calculates distances between centroids and between individuals}

\item{n.for.uniform}{default = 0, if 0 then the CURRENT amount is taken (ie - the uniform subpopulation is spread over the entire polygon),
if nonzero the value itself is taken.}

\item{initial_div}{divisor of expected distance of individuals from the centroid of their clump in the clumped sub-populations.}

\item{max_pct_to_use}{the maximum percentage of clumping to start from - for direct usage of this function this would normally be 100\%,}

\item{percent_random_rise}{the drops in percent that set the levels of clumpedness - like above, this sets the sensitivity of the}

\item{simulations_per_percent}{the amount of simulated populations generated at each level of clumpedness at initial_n, and therefore}

\item{recalculate_poly}{default=T, should the polygons be recalculated to be the polygon of the remaining population.
Use this to encapsulate tagging with an unknown real potential population.}

\item{recalculate_poly_with_rect}{default=F, should the recalcualted polygon (if T for above) should be a bounding rectangle, or the convex polygon.}

\item{plotResults}{as the name implies, whether to plot and display one simulated population for each level of clumpedness using ggplot.}

\item{n_cores}{Default=0, number of cores to use for parallel processing. A value of zero will use the benchmark of parallel:detectCores(logical=FALSE)-1.}

\item{delt_n}{the amount of individuals to remove from each of the simulated population for each step. Higher amount lowers sensitivity,}
}
\value{
a dataframe with length(function_names) + 2 columns and
\deqn{numRows = simulationsPerPercent \times (1+{\frac{initialPct}{percentRandomRise}})}
rows, such that the first two columns css and dfr are the number of individuals in the population and the percent clumpedness of one population,
and the remaining columns hold the output of the function function_names_i for the population.
}
\description{
Using simulate_population_remove_from_original_for_smaller_pops, generate a Large Random-Like Population Distribution (LRPD),
which is later compared to values from the same function. The reasoning here is that:
\itemize{
\item{}{A large enough population (default n=100) such that any basic Clumping Method (CM) will be able to differentiate between a very clumped and very random population}
\item{}{A low Degree of Clumping (dfr) - such that these large simulated populations should come out on the random end (hence "Random-Like"). Default is 0-20\% dfr.}
}
}
