% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{generate_random_population}
\alias{generate_random_population}
\title{generate_random_population}
\usage{
generate_random_population(n, polys)
}
\arguments{
\item{n}{number of individuals in the population}

\item{polys}{list of sf polygons in which to randomize the population}
}
\value{
A dataframe of the random population, with columns TAG (numeric, 1-n), X, and Y.
}
\description{
Generate a random population of size n within a list of sf polygons.
}
