% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{generate_simulation_polygons}
\alias{generate_simulation_polygons}
\title{generate_simulation_polygons}
\usage{
generate_simulation_polygons(
  individual.locations.data.frame,
  use_bounding_rect = T
)
}
\arguments{
\item{individual.locations.data.frame}{Dataframe of individual locations, expecting columns X and Y}

\item{use_bounding_rect}{Default TRUE=T, whether to return an chull/MCP of the locations, or a bounding rectangle of this mcp.}
}
\value{
A list of sf polygons with length 1, which can either be the chull/MCP of individual.locations.data.frame if use_bounding_rect=FALSE,
or a bounding rectangle of said chull/MCP if use_bounding_rect=TRUE.
}
\description{
Given a dataframe of individual locations, calculate and return a list of length 1
with the convexhull (chull, otherwise known as a minimum convex polygon or MCP)
of the points as an sf polygon. If use_bounding_rect == TRUE, return a bounding rectangle of the convex hull.
}
