% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{generate_uniform_population}
\alias{generate_uniform_population}
\title{generate_uniform_population}
\usage{
generate_uniform_population(n, polygons, n.for.dist = 0, noise.ratio = 0.1)
}
\arguments{
\item{n}{number of individuals in the population}

\item{polygons}{list of sf polygons in which to randomize the population}

\item{noise.ratio}{Default 0.1, distance of noise (multiplied by expected distance UniformDist above) to add to each point (so it wont be an exact checkerboard)}

\item{max_dist_from_centroid_divisor}{divisor of the expected distance for distances of each point from it's randomized centroid.
Higher number causes smaller distances between points in the same clump.}

\item{min_dist_between_clumps_multiplier}{the multiplier of the same distance for the distances between the centroids/clumps.}
}
\value{
A dataframe of the clumped population, with columns TAG (numeric, 1-n), X, Y, and clump.id (running index of the clump of the individual).
}
\description{
Generate a uniform/regular population of n points within polygons in polys, return as a dataframe.
Distance between individuals will be \deqn{UniformDist = \sqrt{\frac{AreaOfPolygons}{n}} \pm noise.ratio * \sqrt{\frac{AreaOfPolygons}{n}}  }
}
