% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_pretty_name}
\alias{get_pretty_name}
\title{get_pretty_name}
\usage{
get_pretty_name(col, wrap = F, non.rand.percent.grep = "Rejection Rate")
}
\arguments{
\item{col}{the column name to prettify}

\item{wrap}{Default=F, if T add a line break (backslash n) every 15 characters (at the next space).}

\item{non.rand.percent.grep}{what to replace .non.random.percent with in the pretty name}
}
\value{
the pretty name
}
\description{
Gets pretty name for the generated simulation columns - delimited by spaces instead of periods, and properly capitalized.
}
