% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_special_theme}
\alias{get_special_theme}
\title{get_special_theme}
\usage{
get_special_theme(lockAspectRatio = TRUE)
}
\arguments{
\item{lockAspectRatio}{DEFAULT=TRUE. If TRUE, lock square aspect ratio.}
}
\value{
the ggplot2::theme
}
\description{
Gets a special ggplot theme with a square aspect ratio and larger fonts.
}
