% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap_control_plots}
\alias{heatmap_control_plots}
\title{heatmap_control_plots}
\usage{
heatmap_control_plots(
  heatmap.data,
  lrpd,
  wanted_n = 0,
  just_hist = F,
  with_mnrt = F,
  mnrt.quantile = 0.95
)
}
\arguments{
\item{heatmap.data}{dataframe returned from simulate_population_remove_from_original_for_smaller_pops, with columns per CM.}

\item{lrpd}{dataframe returned from generate_large_randomlike_population_distribution, with columns per CM.}

\item{wanted_n}{wanted population size on which to display the plots. Default 0 dictates to display for the largest size (Initial Sample Size, ISS)}

\item{just_hist}{Default=FALSE. If FALSE, display for each CM a histogram of the LRPD values, with a dashed blue line for the MNRT (95th percentile), and vertical lines of mean CM value from heatmap.data for each dfr.
If TRUE, display from each CM a histogram of the LRPD values, with a dashed blue line for the MNRT (95th percentile), such that values above it are red and values below are green.}

\item{with_mnrt}{Default=FALSE. If TRUE and just_hist = TRUE, mnrt will be drawn as a dashed blue vertical line, and histogram values above will be red.
else if FALSE and just_hist = TRUE, mnrt will not be drawn, and hist vals will be green only.
If just_hist = FALSE, do nothing.}
}
\value{
list of the ggplot2 plots
}
\description{
Outputs plots (via ggplot2) for each Clumping Method (CM), either comparing values of heatmap.data
(simulations resulting from simulate_population_remove_from_original_for_smaller_pops) to the
Large Random-Like Population Distributions (LRPD, result of generate_large_randomlike_population_distribution), or
Only displaying the LRPD with the Minimum Non-Random Threshold (MNRT, the 95th percentile of the LRPD) for simplicity.
}
