% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{plot_population}
\alias{plot_population}
\title{plot_population}
\usage{
plot_population(
  polygons,
  population,
  pct_random,
  curr_n,
  dot_size = 2,
  wanted.aspect.ratio = 1,
  draw.poly = TRUE,
  with.text = TRUE,
  colormap = NULL
)
}
\arguments{
\item{polygons}{list of sf polygons - drawn as lines on the plot}

\item{population}{results dataframe from any of the generating functions (clumped, random, uniform). Has columns X, Y, and is.random (is the point of a random subpopulation)}

\item{pct_random}{Percent of the population that is random - for the plot title}

\item{curr_n}{Number of individuals - for the plot title}

\item{dot_size}{Default=2, size of the dots on the plot}

\item{wanted.aspect.ratio}{Default=1, wanted aspect ratio of the plot}

\item{draw.poly}{Default=TRUE, should polygon be drawn around points}

\item{with.text}{Default=TRUE, should text (title, legend, axes) appear or not}

\item{coloramp}{optional colormap as a named vector with random, clumped, and uniform. If NULL, default is used
c(random='blue', clumped='red', uniform='darkgreen')}
}
\value{
the ggplot2 plot
}
\description{
Plot a simulated population. Plot has polygons drawn as lines, with the individuals from population as red dots if is.random=FALSE, blue triangles if is.random=TRUE.
}
