% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_simulation_data}
\alias{plot_simulation_data}
\title{plot_simulation_data}
\usage{
plot_simulation_data(
  aggregated.nrp.heatmap.data,
  wanted_ns = c(),
  wanted_ds = c(),
  use.nrp = TRUE,
  doHeatmaps = TRUE,
  doLinePlots = FALSE,
  doFlippedLinePlots = FALSE,
  linewidth.lineplot = 1,
  lineplot.min.color = "blue",
  lineplot.max.color = "red",
  AOI = NULL,
  linewidth.for.AOI = 1,
  fname.in.title = TRUE,
  grep.pretty.name.non.random.percent = "Rejection Rate"
)
}
\arguments{
\item{aggregated.nrp.heatmap.data}{Dataframe result from calculate_nonrandom_percent_on_heatmap_data}

\item{wanted_ns}{relevant for doLinePlots==TRUE and/or doFlippedLinePlots==TRUE. If non-empty, only these values of Current Sample Size (css) will be displayed in these plots.
If empty (default), specific values of CSS will be used, as the following percentiles: 15, 25, 50, 75, 85, 100.}

\item{wanted_ds}{relevant for doLinePlots==TRUE and/or doFlippedLinePlots==TRUE. If non-empty, only these values of Degrees Of Clumping (dfr) will be displayed in these plots.
If empty (default), all levels will be displayed.}

\item{use.nrp}{Default=TRUE, if TRUE the Non-Random Percents (ratio of Simulation Repetitions which acheived a value over the 95th percentile of the Large Random-Like Population Distribution)
for each CM are the ones used for the plots. Otherwise the numeric value of the CM is what is used.}

\item{doHeatmaps}{Default=TRUE, output a heatmap for each CM. The Fill (yellow to red) is the NRP or the mean CM value (depending on use.nrp), the y-axis is the levels of dfr, and the x-axis is the levels of CSS.}

\item{doLinePlots}{Default=FALSE, output a dfr-Comparison line plot for each CM. The y-axis is the NRP or the mean CM value (depending on use.nrp), the the x-axis is the levels of CSS, and there is a line for each level of dfr.}

\item{doFlippedLinePlots}{Default=FALSE, output a CSS-Comparison line plot for each CM. The y-axis is the NRP or the mean CM value (depending on use.nrp), the the x-axis is the levels of dfr, and there is a line for each level of CSS in wanted_ns}

\item{linewidth.lineplot}{DEFAULT=1, linewidth for both regular and flipped lineplots.}

\item{lineplot.min.color}{Default='blue', for min of gradient scaling of the colors of the lines.}

\item{lineplot.max.color}{Default='red', for max of gradient scaling of the colors of the lines.}

\item{AOI}{Default=NULL, if not-null, adds a rectangle around the AOI on the heatmap. Order must be:
\itemize{
\item min.css
\item min.dfr
\item max.css
\item max.dfr
}

Note that it is safeguarded for order, and also that minimi and maximi are trimmed to the max possible value.}

\item{linewidth.for.AOI}{Default=1, if AOI is not null, the width of the rectangle of the AOI.}

\item{fname.in.title}{Default=TRUE, if to display the pretty name of the quantification method as the main title or appended to the proper axis.}

\item{grep.pretty.name.non.random.percent}{pretty replacement for non.random.percent in plots}
}
\value{
list of the ggplot2 plots
}
\description{
Outputs various plots (via ggplot2) for each Clumping Method (CM) in the columns of aggregated.nrp.heatmap.data, with the Degrees of Clumpedness (dfr) vs. the Current Sample Size (CSS).
}
