% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots_of_example_data}
\alias{plots_of_example_data}
\title{plots_of_example_data}
\usage{
plots_of_example_data(
  example.data.with.values,
  lrpd,
  cutoffs,
  withSimData = F,
  withGradient = F,
  numericCutoffsForGradient = c(-2, -1),
  line.width.gradient = 2,
  hjust.labels = 0.5,
  vjust.labels = 2,
  grep.pretty.name.non.random.percent = "Rejection Rate"
)
}
\arguments{
\item{example.data.with.values}{Dataframe outputted by calculate_values_example_data}

\item{lrpd}{Dataframe outputted by generate_large_randomlike_population_distribution}

\item{cutoffs}{a vector with the same length as the amount of CMs in the example data. This is so that only populations whose size
permits usage of the CM (after checking to see at which population size this is not relevant with the plots of
plot_simulation_data). so every population with n<cutoffs_i will NOT appear in the plot of CM_i=example.data.with.values_(i+2) (since the first two columns are POPULATION and CSS)}

\item{withSimData}{Default=FALSE. If TRUE, for each CM the values of populations relevant for the cutoff of each CM are displayed as vertical lines, overlaid on a histogram
of the LRPD values with the Minimum Non-Random Threshold (95th percentile of the LRPD values) marked in a dashed blue line.
If FALSE, for each CM the values of the population are displayed in a dot plot, where the X axis is simply sorted according to the CM value on the y axis. The dots are sized according to the CSS of the population, and a shorthand of the first letter of each word in POPULATION is placed next to each dot.}

\item{withGradient}{default=FALSE. If withSimData=TRUE, it is ignored. If withGradient=FALSE, the plot as described above is displayed per CM.
If withGradient=TRUE, it is overlain on a yellow-to-red gradient, and the value of numericCutoffForGradient comes into play, such that populations
with a CM value > numericCutoffForGradient will be colored purple, and those under green.}

\item{line.width.gradient}{default=2, for gradient display, width of the vertical lines on the gradient.}

\item{hjust.labels}{default=0.5, hjust for labels of population names on the points}

\item{vjust.labels}{default=2, vjust for labels of population names on the points}

\item{grep.pretty.name.non.random.percent}{pretty replacement for non.random.percent in legends}

\item{numericCutoffForGradient}{default=-1.7, which is the relevant cutoff for clark.evans.naive on the supplied bird.location.data. If withSimData=TRUE or withGradient=FALSE, it is ignored.
If withSimData=FALSE and withGradient=TRUE, the aforementioned coloration is done, and the value appears as a black horizontal line.}
}
\value{
list of the ggplot2 plots
}
\description{
given example data in example.data.with.values with a column for each Clumping Method (CM), output comparative plots for each CM, taking into account the Currrent Sample Size (CSS) of each population to see if relevant for the CM.
}
