% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{population_type}
\alias{population_type}
\title{population_type}
\usage{
population_type
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   possible population type names - clumped, random, uniform, and subrandom.
}\if{html}{\out{</div>}}
}
