% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{simulate_population_remove_from_original_for_smaller_pops}
\alias{simulate_population_remove_from_original_for_smaller_pops}
\title{simulate_population_remove_from_original_for_smaller_pops}
\usage{
simulate_population_remove_from_original_for_smaller_pops(
  polygons,
  function_names,
  helper_functions_to_export,
  sim.type = simulation_type$random.to.clumped,
  uniform.random.generation.type =
    uniform_random_generation_type$split.polygon.and.generate.separately,
  initial_n = 20,
  n.for.calc = 0,
  n.for.clumping = 0,
  n.for.uniform = 0,
  initial_div = 2,
  n_clumps = 0,
  uniform.noise.ratio = 0.1,
  delt_n = 5,
  initial_pct = 100,
  percent_random_rise = 20,
  simulations_per_percent = 20,
  recalculate_poly = T,
  recalculate_poly_with_rect = F,
  plotResults = FALSE,
  ns_to_plot = c(),
  n_cores = 0
)
}
\arguments{
\item{polygons}{list of sf polygons where the populations can be generated.}

\item{function_names}{vector of function names (Clumping Methods, CMs), which are first tested by check_function_validity.}

\item{helper_functions_to_export}{vector of additional functions needed for the CMs to run. Tested only for existence in memory.}

\item{sim.type}{default = simulation_type$random.to.clumped, which part of the range to simulate.}

\item{uniform.random.generation.type}{default = uniform_random_generation_type$split.polygon.and.generate.separately, how to generate the populations in the uniform-random range (DFR -100 to 0), if relevant.}

\item{initial_n}{number of individual of each original population Initial Sample Size (ISS) in the paper.}

\item{n.for.calc}{sent into each of the function in function_names, but does not have to be used by each of these. Is meant}

\item{n.for.clumping}{similarly, the clumping function above calculates distances between centroids and between individuals}

\item{n.for.uniform}{default = 0, if 0 then the CURRENT amount is taken (ie - the uniform subpopulation is spread over the entire polygon),
if nonzero the value itself is taken.}

\item{initial_div}{divisor of expected distance of individuals from the centroid of their clump in the clumped sub-populations.}

\item{n_clumps}{number of clumps wanted in the clumped populations. Default=0 uses the default function of floor(sqrt(n))}

\item{uniform.noise.ratio}{default = 0.1, noise for uniform function}

\item{delt_n}{the amount of individuals to remove from each of the simulated population for each step. Higher amount lowers sensitivity,}

\item{initial_pct}{the maximum percentage of clumping to start from - for direct usage of this function this would normally be 100\%,}

\item{percent_random_rise}{the drops in percent that set the levels of clumpedness - like above, this sets the sensitivity of the}

\item{simulations_per_percent}{the amount of simulated populations generated at each level of clumpedness at initial_n, and therefore}

\item{recalculate_poly}{default=T, should the polygons be recalculated to be the polygon of the remaining population.
Use this to encapsulate tagging with an unknown real potential population.}

\item{recalculate_poly_with_rect}{default=F, should the recalcualted polygon (if T for above) should be a bounding rectangle, or the convex polygon.}

\item{plotResults}{as the name implies, whether to plot and display one simulated population for each level of clumpedness using ggplot.}

\item{ns_to_plot}{which ns should be plotted. If empty, will be only the original population size (ISS)}

\item{n_cores}{Default=0, number of cores to use for parallel processing. A value of zero will use the benchmark of parallel:detectCores(logical=FALSE)-1.
Results of benchmark tests with initial_n = 30, simulations_per_percent = 100, delt_n = 5, and percent_random_rise = 20
are below - all values over 7 cores took less than a minute.

2 cores \tab 1.158628 mins \tab 1.431186 mins \tab 1.447352 mins\cr
3 cores \tab 1.450939 mins \tab 1.615264 mins \tab 1.551022 mins\cr
4 cores \tab 1.267707 mins \tab 1.110228 mins \tab 1.097605 mins\cr
5 cores \tab 1.016974 mins \tab 1.0203 mins \tab 1.021301 mins\cr
6 cores \tab 1.29498 mins \tab 1.368751 mins \tab 1.404914 mins\cr
7 cores \tab 1.263218 mins \tab 1.3216 mins \tab 59.86292 secs\cr
8 cores \tab 59.03853 secs \tab 59.46413 secs \tab 57.9628 secs\cr
9 cores \tab 59.18239 secs \tab 57.93962 secs \tab 58.88627 secs\cr
10 cores \tab 57.72106 secs \tab 1.039918 mins \tab 57.46983 secs\cr
11 cores \tab 57.00588 secs \tab 57.55166 secs \tab 58.5118 secs\cr
12 cores \tab 57.31099 secs \tab 1.042125 mins \tab 55.91966 secs\cr
13 cores \tab 56.57559 secs \tab 57.63825 secs \tab 1.015002 mins\cr
14 cores \tab 58.69229 secs \tab 57.11787 secs \tab 58.8794 secs\cr
15 cores \tab 59.30693 secs \tab 58.96996 secs \tab 59.86175 secs\cr
16 cores \tab 59.50359 secs \tab 59.90103 secs \tab 59.55054 secs}
}
\value{
a dataframe with length(function_names) + 2 columns and
\deqn{numRows = simulationsPerPercent \times {\frac{initialN}{deltN}} \times (1+{\frac{initialPct}{percentRandomRise}})}
rows, such that the first two columns css and dfr are the number of individuals in the population and the percent clumpedness of one population,
and the remaining columns hold the output of the function function_names_i for the population.
}
\description{
Core function of the process, does  both the heatmap simulation data and the null-model distribution data (via the wrapper function below).
The general flow is as follows:
\itemize{
\item{} {Generate simulations_per_percent populations of size initial_n of each level of initial_pct/percent_random_rise + 1, such that each such
simulated population is X\% clumped and (100-X)\% random. Calculate all the functions in function_names on the populations.}
\item{} {Until the populations dwindle to minimum size 2, remove delt_n individuals from each population at each step, and recalculate
all the functions in function_names.}
}
This is meant to test the robustness of the functions in function_names of differentiating between different levels of
clumpedness for even small subpopulations, where it is assumed that only some of the real population have been geotagged.
Note that this function parallely runs on N-1 out of N cores of the PC, and that the heavy runtime parameter is set by a mix of
percent_random_rise and simulations_per_percent, since these set the original number of populations that need to be simulated.
}
