% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{simulation_type}
\alias{simulation_type}
\title{simulation_type}
\usage{
simulation_type
}
\description{
Possible simulation type names:
\itemize{
\item{random.to.clumped}: {Population in the half-range of random to clumped (DFR 0 to 100)}
\item{uniform.to.random}: {Population in the half-range of uniform to random (DFR -100 to 0)}
\item{uniform.to.random.to.clumped}: {Population in the full range of uniform to random to clumped (DFR -100 to 100)}
}
}
