% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{split_poly}
\alias{split_poly}
\title{split_poly}
\usage{
split_poly(poly, ratio.first.poly)
}
\arguments{
\item{poly}{Original polygon as an sf polygon}

\item{ratio.first.poly}{0-1, the approximate percentage to give the first of the two polygons.}
}
\value{
A list of five parameters:
\itemize{
\item{ratio.first.poly}: {same as input, for tracking purposes}
\item{full.poly}: {same as input polygon, for tracking purposes}
\item{poly.1}: {a sub-polygon of approximately ratio.first.poly of the large polygon, or NULL if ratio.first.polygon==0. Will be at the bottom if h>w in the original polygon,
and at the left if h<=w in the original polygon}
\item{poly.2}: {a sub-polygon of approximately 1-ratio.first.poly of the large polygon, or NULL if ratio.first.polygon==1. Will be at the top if h>w in the original polygon,
and at the right if h<=w in the original polygon}
\item{plot}: {plot of the algorithmic steps, if wanted.}
}
}
\description{
Split polygon into 2 APPROXIMATELY equal sized polygons, s.t the first will be approximately ratio.first.poly of the bottom (if h>w) or left (if h<=w) of the polygon,
and the second will be the remaining (1-ratio.first.poly) of the top (if h>w) or right (if h<=w) of the polygon.
In case of ratio.first.poly==0 the first polygon will be null and the second the full polygon,
and vice versa if ratio.first.poly==1.
Note that currently the split is according to the median of the longer of the the two axes of the polygon,
and therefore not necessarily will the area be equal to ratio.first.poly or 1-ratio.first.poly. However, this is inconsequential to the calculations afterwards, since
the uniform subpopulation will still be highly uniform within its subpolygon.
}
