% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSpread.R
\name{uniform_random_generation_type}
\alias{uniform_random_generation_type}
\title{uniform_random_generation_type}
\usage{
uniform_random_generation_type
}
\description{
Possible ways to generate the random-uniform mixed populations:
\itemize{
\item{generate.mix.in.full.polygon}: {The populations are generated in the same way as the random-clumped mixed populations -
CSS x abs(DFR) uniform points within the full polygon, and CSS x (1 - abs(DFR)) random points within the full polygon.}
\item{split.polygon.and.generate.separately}: {The populations are generated in different subpolygons that make up the full polygon,
by first splitting the polygon into two approximately equal sized polygons of DFR and (1-DFR) ratios of the full polygon.
Then, the uniform subpopulation is generated within the DFR sub-polygon, and the random subpopulation within the (1-DFR) subpolygon.
This generation in our opinion is more correct, as it allows the uniform sub-population to retain their distance, and not have the
random additions encroach the uniformity.}
}
}
